feat: add some refactor
