+++ src/main/java/com/rakovets/course/java/core/practice/concurrency/producer_queue_consumer/exception/UserInputException.java
@@ -1,6 +1,8 @@
 package com.rakovets.course.java.core.practice.concurrency.producer_queue_consumer.exception;
 
-public class UserInputException extends Exception {
+import java.util.InputMismatchException;
+
+public class UserInputException extends InputMismatchException {
     public UserInputException(String message) {
         super(message);
     }
