+++ src/main/java/com/rakovets/course/java/core/practice/concurrency/producer_queue_consumer/subscriber/Consumer.java
@@ -1,20 +1,20 @@
 package com.rakovets.course.java.core.practice.concurrency.producer_queue_consumer.subscriber;
 
-import com.rakovets.course.java.core.practice.concurrency.producer_queue_consumer.utils.CreateFile;
+import com.rakovets.course.java.core.practice.concurrency.producer_queue_consumer.utils.FileCreator;
 
 import java.util.Queue;
 import java.util.concurrent.TimeUnit;
 
-public class ConsumerImpl implements Runnable, ISleep {
+public class Consumer implements Runnable, ISleep {
     private final Queue<Integer> queue;
 
-    public ConsumerImpl(Queue<Integer> queue) {
+    public Consumer(Queue<Integer> queue) {
         this.queue = queue;
     }
 
     @Override
     public void run() {
-        CreateFile newFile = new CreateFile("output.txt");
+        FileCreator newFile = new FileCreator("output.txt");
         System.out.println("Working...");
 
         while (queue.peek() != null) {
